Organizer of Robotics Workshop
We have organized a robotics workshop. The Robotics Workshop was a great opportunity for us to gain hands-on experience in designing and programming robots. We learned about the different components of robotics, including sensors, controllers, and software. This workshop was designed to introduce you to the fascinating
world of robotics, such as line-following robots.